requirements.txt
imageio
psutil
flask
requests
json5
numpy>=2.3.3
# no need for api
# uiautomation>=2
celery>=5.3
python-dotenv>=1.0.0
# Production WSGI server (Windows and Linux compatible)
waitress>=3.0.0
# Optional broker dependencies (choose one)
redis>=7.0.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9.11
alembic>=1.17
cos-python-sdk-v5>=1.9.38